Run Tornado server as a module
Both the client and the gRPC server are run as modules:

* Running the client is described in the README as `python -m
  hbi.client`.
* Running the gRPC server is in the serve.py scriptlet as `python -m
  hbi.server.grpc_server`.

Replaced serve_rest.py scriptlet with a possibility to run the
REST (Tornado) server the same way as `python -m
hbi.server.tornado_server`. The scriptlet did nothing more than just
calling a single function. This way it will be possible to use the
same Watchdog machinery as the gRPC server does.